pyproject.toml
# https://peps.python.org/pep-0517/
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

# https://peps.python.org/pep-0621/
[project]
name = "pymmcore-plus"
description = "pymmcore superset providing improved APIs, event handling, and a pure python acquisition engine"
keywords = ["microscope", "micro-manager", "smart-microscopy"]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "BSD 3-Clause License" }
authors = [
    { name = "Talley Lambert", email = "talley.lambert@gmail.com" },
    { name = "Federico Gasparoli", email = "federico.gasparoli@gmail.com" },
    { name = "Ian Hunt-Isaak", email = "ianhuntisaak@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Hardware",
    "Topic :: System :: Hardware :: Hardware Drivers",
    "Topic :: Utilities",
]
dynamic = ["version"]
dependencies = [
    "numpy >=1.25.2",
    "numpy >=1.26.0; python_version >= '3.12'",
    "numpy >=2.1.0; python_version >= '3.13'",
    "ome-types >=0.6.0",
    "ome-writers[acquire-zarr,tensorstore,tifffile]>=0.0.1b0",
    "platformdirs >=3.0.0",
    "psygnal >=0.10",
    "pymmcore >=11.10.0.74.0",
    "tensorstore >=0.1.67,!=0.1.72",
    "tensorstore >=0.1.71,!=0.1.72; python_version >= '3.13'",
    "typing-extensions >=4", # not actually required at runtime
    "useq-schema >=0.7.2",
    # cli requirements included by default for now
    "typer >=0.4.2",
    "rich >=10.2.0",
]

# extras
# https://peps.python.org/pep-0621/#dependencies-optional-dependencies
[project.optional-dependencies]
cli = ["typer >=0.4.2", "rich >=10.2.0"]
io = ["tifffile >=2021.6.14", "zarr >=2.15,<3"]
PySide2 = ["PySide2 >=5.15.2.1"]
PySide6 = ["PySide6 ==6.7.3"]
PyQt5 = ["PyQt5 >=5.15.4"]
PyQt6 = ["PyQt6 >=6.4.2"]

[dependency-groups]
nano = ["pymmcore-nano==11.10.0.74.0"]
docs = [
    "mkdocs >=1.4",
    "mkdocs-material>=9.5",
    "mkdocstrings ==0.22.0",
    "mkdocs-autorefs ==1.3.1",
    "mkdocstrings-python ==1.1.2",
    "mkdocs-typer ==0.0.3",
]
test = [
    "pymmcore-plus[io]",
    "msgspec >= 0.19",
    "msgpack >=1",
    "pytest-cov >=7",
    "ipython>=8.18.0",
    "pytest >=8",
    "xarray >=2024.1",
    "lxml >=6.0",
]
test-codspeed = [{ include-group = "test" }, "pytest-codspeed >=3.2.0"]
test-qt = [{ include-group = 'test' }, "pytest-qt ==4.4", "qtpy >=2"]
PyQt6 = [{ include-group = 'test-qt' }, "pymmcore-plus[PyQt6]"]
PySide6 = [{ include-group = 'test-qt' }, "pymmcore-plus[PySide6]"]
PyQt5 = [{ include-group = 'test-qt' }, "pymmcore-plus[PyQt5]"]
PySide2 = [{ include-group = 'test-qt' }, "pymmcore-plus[PySide2]"]
dev = [
    # { include-group = "docs" },
    { include-group = "PyQt6" },
    "pdbpp>=0.11.6 ; sys_platform != 'win32'",
    "mypy>=1.14.1",
    "pre-commit>=4.1.0",
    "ruff>=0.9.4",
    "pydantic >2.7.4; python_version >= '3.13'",
]

[tool.uv.sources]
pymmcore-plus = { workspace = true }

[tool.uv]
override-dependencies = [
    "griffe @ git+https://github.com/tlambert03/griffe@recursion#egg=griffe",
]

[project.urls]
Source = "https://github.com/pymmcore-plus/pymmcore-plus"
Tracker = "https://github.com/pymmcore-plus/pymmcore-plus/issues"
Documentation = "https://pymmcore-plus.github.io/pymmcore-plus"

[project.scripts]
mmcore = "pymmcore_plus._cli:main"

# https://hatch.pypa.io/latest/config/metadata/
[tool.hatch.version]
source = "vcs"

[tool.hatch.metadata]
allow-direct-references = true

# https://hatch.pypa.io/latest/config/build/#file-selection
[tool.hatch.build.targets.sdist]
include = ["/src", "/tests"]

[tool.hatch.build.targets.wheel]
only-include = ["src"]
sources = ["src"]

# https://docs.astral.sh/ruff/rules/
[tool.ruff]
line-length = 88
target-version = "py39"
fix = true
unsafe-fixes = true

[tool.ruff.lint]
pydocstyle = { convention = "numpy" }
select = [
    "E",    # style errors
    "F",    # flakes
    "W",    # warnings
    "D",    # pydocstyle
    "I",    # isort
    "UP",   # pyupgrade
    "C4",   # flake8-comprehensions
    "B",    # flake8-bugbear
    "A001", # flake8-builtins
    "RUF",  # ruff-specific rules
    "TID",  # tidy
    "TC",   # typecheck
    "SLF",  # private-access
]
ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D401", # First line should be in imperative mood (remove to opt in)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "SLF"]
"examples/*.py" = ["D"]
"_cli.py" = ["B008"]
"docs/*.py" = ["A", "D"]

# https://docs.astral.sh/ruff/formatter/
[tool.ruff.format]
docstring-code-format = true

# https://docs.pytest.org/en/6.2.x/customize.html
[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore:Failed to disconnect::pytestqt",
    "ignore:numpy.core.multiarray is deprecated",
    "ignore:Focus direction is unknown",
    "ignore:'BaseCommand' is deprecated:DeprecationWarning:typer",
    "ignore:Failing to pass a value to the 'type_params' parameter",
]
markers = ["run_last: mark a test to run last"]

# https://mypy.readthedocs.io/en/stable/config_file.html
[tool.mypy]
files = "src/**"
strict = true
disallow_any_generics = false
disallow_subclassing_any = false
show_error_codes = true
pretty = true
plugins = "pydantic.mypy"
# see https://github.com/python/mypy/issues/5374 and related discussions
# it causes more pain than it solves
disable_error_code = ["type-abstract"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# https://coverage.readthedocs.io/en/6.4/config.html
[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@overload",
    "except ImportError",
    "except Exception as e:",     # ok not to cover bare exception catches in tests
    "raise AssertionError",
    "\\.\\.\\.",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
show_missing = true

[tool.coverage.run]
source = ['pymmcore_plus']
omit = ["src/pymmcore_plus/_build.py"]

# https://github.com/mgedmin/check-manifest#configuration
[tool.check-manifest]
ignore = [
    ".github_changelog_generator",
    ".pre-commit-config.yaml",
    ".ruff_cache/**/*",
    "tests/**/*",
    "tox.ini",
]

[tool.typos.default]
extend-ignore-identifiers-re = ["(?i)nd2?.*", "(?i)ome", "anager", "ba"]
